pyproject.toml
[tool.poetry]
name = "homework-24-1"
version = "0.1.0"
description = ""
authors = ["Oleg Usachev <usaechv97@mail.ru>"]
readme = "README.md"
packages = [{include = "homework_24"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
psycopg2 = "^2.9.9"
pillow = "^10.2.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.4.2"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
stripe = "^8.4.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
